// src/app/interfaces/reg-form-usuario.interface.ts
export interface RegUsuarioForm{
    userNombres: string;
    userApellidos: string;
    userEmail: string;
    userEstado: boolean;
    userRolID: string;
    userDateAdd: string;
    userContacto: string;
    userPassword: string;
    userConfirPassword: string;
    userSobreMi: string;
    userAvatar: string;
}